Use DotNotZip for folder creation inside the ZIP
Added DotNotZip to make it easy to create folders inside the archive
Added ability to loop through entire dod_downloads folder looking for BSPs
Added sprites/mapsprites//fexplo.spr to dod exclusion list as this is in dod_merderet.bsp erroneously
Added RelativePath to AssociatedFile to be used in the creation of folders in the archive
Fixed bug where only GameDirectories[0] was searched for the bsp file
All files loaded into memory from the RESGen.exe output are switched to all lowercase
All files loaded into memory from the exclusion lists are switched to all lowercase
Ambient sounds frenchcountry.wav, moo.wav, and opera.wav are included in included maps, but not in the exclusions list. This needs researched.